renamings and docstrings
